TRAD-13 Implement service to get onboardDate from market_info by symbol

+++ src/main/java/com/tradey/technical_analysis/domain/services/SymbolInfoService.java
@@ -4,13 +4,11 @@ import com.tradey.technical_analysis.domain.entity.SymbolInfoEntity;
 import com.tradey.technical_analysis.domain.repositories.SymbolInfoRepository;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigInteger;
-
 @RequiredArgsConstructor
 public class SymbolInfoService {
     private final SymbolInfoRepository symbolInfoRepository;
 
-    public BigInteger getOnboardDate(String exchangeType, String symbol) {
+    public Long getOnboardDate(String exchangeType, String symbol) {
         SymbolInfoEntity entity = symbolInfoRepository.get(exchangeType, symbol);
         return entity == null ? null : entity.getOnboardDate();
     }
