+++ src/main/java/com/tradey/technical_analysis/infrastructure/dto/SymbolInfoDTO.java
@@ -4,13 +4,11 @@ import com.tradey.technical_analysis.domain.entity.SymbolInfoEntity;
 
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigInteger;
-
 @RequiredArgsConstructor
 public class SymbolInfoDTO {
     private final String exchange_type;
     private final String symbol;
-    private final BigInteger onboard_date;
+    private final long onboard_date;
 
     public SymbolInfoEntity toEntity() {
         return new SymbolInfoEntity(exchange_type, symbol, onboard_date);
